Make State generic: move state outside of service
