+++ server/lib.rs
@@ -14,3 +14,4 @@ pub mod parameters;
 pub mod server;
 pub mod service;
 mod util;
+pub(crate) mod state;
