+++ server/service/http/typedb_service.rs
@@ -6,7 +6,7 @@
 
 use std::{collections::HashMap, net::SocketAddr, sync::Arc, time::Duration};
 
-use crate::service::state::ServerState;
+use crate::state::ServerState;
 use crate::{
     authentication::Accessor,
     service::{
