+++ server/service/mod.rs
@@ -7,7 +7,6 @@ use serde::{Deserialize, Serialize};
 
 pub(crate) mod grpc;
 pub mod http;
-pub(crate) mod state;
 mod transaction_service;
 
 #[derive(Serialize, Deserialize, Debug, Clone, Copy, PartialOrd, PartialEq, Eq, Hash)]
